backend/requirements.txt
fastapi
uvicorn
anthropic
python-multipart
jinja2
streamlit
python-dotenv
groq
slowapi
requests